requirements.txt
feature_engine==1.9.3
kaggle==1.7.4.5
matplotlib==3.10.7
mlflow==3.5.1
mlflow_skinny==3.5.1
mlflow_tracing==3.5.1
pandas==2.3.3
python-dotenv==1.1.1
scikit_learn==1.7.2
seaborn==0.13.2
SQLAlchemy==2.0.44
tqdm==4.67.1